- extract in new files to ease up the integration process
