+++ src/main/java/com/thebuzzmedia/exiftool/ExifTool.java
@@ -15,29 +15,21 @@
  */
 package com.thebuzzmedia.exiftool;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
 import java.lang.reflect.Array;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
-import java.util.EnumMap;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.StringTokenizer;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.TreeMap;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.regex.Pattern;
 
 import org.slf4j.Logger;
@@ -271,18 +263,17 @@ public class ExifTool {
 	 * <p/>
 	 * Default value is "<code>ExifTool Cleanup Thread</code>".
 	 */
-	private static final String CLEANUP_THREAD_NAME = "ExifTool Cleanup Thread";
+	static final String CLEANUP_THREAD_NAME = "ExifTool Cleanup Thread";
 
 	/**
 	 * Compiled {@link Pattern} of ": " used to split compact output from
 	 * ExifTool evenly into name/value pairs.
 	 */
-	private static final Pattern TAG_VALUE_PATTERN = Pattern
-			.compile("\\s*:\\s*");
-	private static final String STREAM_CLOSED_MESSAGE = "Stream closed";
+	static final Pattern TAG_VALUE_PATTERN = Pattern.compile("\\s*:\\s*");
+	static final String STREAM_CLOSED_MESSAGE = "Stream closed";
 	private static final String EXIF_DATE_FORMAT = "yyyy:MM:dd HH:mm:ss";
 
-	private static Logger log = LoggerFactory.getLogger(ExifTool.class);
+	static Logger log = LoggerFactory.getLogger(ExifTool.class);
 
 	private final Map<Feature, Boolean> featureSupportedMap = new HashMap<Feature, Boolean>();
 	private final Set<Feature> featureEnabledSet = EnumSet
@@ -503,13 +494,6 @@ public class ExifTool {
 	public ExifTool setWriteOptions(WriteOptions options) {
 		defWriteOptions = options;
 		return this;
-	}
-
-	// ================================================================================
-	// OLD API
-	public static enum Format {
-		NUMERIC,
-		HUMAN_READABLE
 	}
 
 	public Map<Tag, String> getImageMeta(File image, Tag... tags)
@@ -843,979 +827,5 @@ public class ExifTool {
 	}
 
 	// ================================================================================
-	/**
-	 * Represents an external exif process. Works for both single use and keep
-	 * alive modes. This is the actual process, with streams for reading and
-	 * writing data.
-	 */
-	public static final class ExifProcess {
-		public static VersionNumber readVersion(String exifCmd) {
-			ExifProcess process = new ExifProcess(false, Arrays.asList(exifCmd,
-					"-ver"));
-			try {
-				return new VersionNumber(process.readLine());
-			} catch (IOException ex) {
-				throw new RuntimeException(String.format(
-						"Unable to check version number of ExifTool: %s",
-						exifCmd));
-			} finally {
-				process.close();
-			}
-		}
 
-		private final ReentrantLock closeLock = new ReentrantLock(false);
-		private final boolean keepAlive;
-		private final Process process;
-		private final BufferedReader reader;
-		private final OutputStreamWriter writer;
-		private volatile boolean closed = false;
-
-		public ExifProcess(boolean keepAlive, List<String> args) {
-			this.keepAlive = keepAlive;
-			log.debug(String
-					.format("Attempting to start ExifTool process using args: %s",
-							args));
-			try {
-				this.process = new ProcessBuilder(args).start();
-				this.reader = new BufferedReader(new InputStreamReader(
-						process.getInputStream()));
-				this.writer = new OutputStreamWriter(process.getOutputStream());
-				log.debug("\tSuccessful");
-			} catch (Exception e) {
-				String message = "Unable to start external ExifTool process using the execution arguments: "
-						+ args
-						+ ". Ensure ExifTool is installed correctly and runs using the command path '"
-						+ args.get(0)
-						+ "' as specified by the 'exiftool.path' system property.";
-
-				log.debug(message);
-				throw new RuntimeException(message, e);
-			}
-		}
-
-		public synchronized Map<String, String> sendToRunning(List<String> args)
-				throws IOException {
-			if (!keepAlive) {
-				throw new IOException("Not KeepAlive Process");
-			}
-			StringBuilder builder = new StringBuilder();
-			for (String arg : args) {
-				builder.append(arg).append("\n");
-			}
-			builder.append("-execute\n");
-			writeFlush(builder.toString());
-			return readResponse();
-		}
-
-		public synchronized void writeFlush(String message) throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			writer.write(message);
-			writer.flush();
-		}
-
-		public synchronized String readLine() throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			return reader.readLine();
-		}
-
-		public synchronized Map<String, String> readResponse()
-				throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			log.debug("Reading response back from ExifTool...");
-			Map<String, String> resultMap = new HashMap<String, String>(500);
-			String line;
-
-			while ((line = reader.readLine()) != null) {
-				if (closed)
-					throw new IOException(STREAM_CLOSED_MESSAGE);
-				String[] pair = TAG_VALUE_PATTERN.split(line, 2);
-
-				if (pair.length == 2) {
-					resultMap.put(pair[0], pair[1]);
-					log.debug(String.format("\tRead Tag [name=%s, value=%s]",
-							pair[0], pair[1]));
-				}
-
-				/*
-				 * When using a persistent ExifTool process, it terminates its
-				 * output to us with a "{ready}" clause on a new line, we need
-				 * to look for it and break from this loop when we see it
-				 * otherwise this process will hang indefinitely blocking on the
-				 * input stream with no data to read.
-				 */
-				if (keepAlive && line.equals("{ready}")) {
-					break;
-				}
-			}
-			return resultMap;
-		}
-
-		public boolean isClosed() {
-			return closed;
-		}
-
-		public void close() {
-			if (!closed) {
-				closeLock.lock();
-				try {
-					if (!closed) {
-						closed = true;
-						try {
-							log.debug("Closing Read stream...");
-							reader.close();
-							log.debug("\tSuccessful");
-						} catch (Exception e) {
-							// no-op, just try to close it.
-						}
-
-						try {
-							log.debug("Attempting to close ExifTool daemon process, issuing '-stay_open\\nFalse\\n' command...");
-							writer.write("-stay_open\nFalse\n");
-							writer.flush();
-						} catch (IOException ex) {
-							// log.error(ex,ex);
-						}
-
-						try {
-							log.debug("Closing Write stream...");
-							writer.close();
-							log.debug("\tSuccessful");
-						} catch (Exception e) {
-							// no-op, just try to close it.
-						}
-
-						log.debug("Read/Write streams successfully closed.");
-
-						try {
-							process.destroy();
-						} catch (Exception e) {
-							//
-						}
-						// process = null;
-
-					}
-				} finally {
-					closeLock.unlock();
-				}
-			}
-		}
-	}
-
-	/**
-	 * A Proxy to an Exif Process, will restart if backing exif process died, or
-	 * run new one on every call.
-	 * 
-	 * @author Matt Gile, msgile
-	 */
-	public interface ExifProxy {
-		public void startup();
-
-		public Map<String, String> execute(long runTimeoutMills,
-				List<String> args) throws IOException;
-
-		public boolean isRunning();
-
-		public void shutdown();
-	}
-
-	/**
-	 * Manages an external exif process in keep alive mode.
-	 */
-	public static class KeepAliveExifProxy implements ExifProxy {
-		private final List<String> startupArgs;
-		private final AtomicBoolean shuttingDown = new AtomicBoolean(false);
-		private final Timer cleanupTimer = new Timer(CLEANUP_THREAD_NAME, true);
-		private long inactivityTimeout = 0;
-		private volatile long lastRunStart = 0;
-		private volatile ExifProcess process;
-
-		public KeepAliveExifProxy(String exifCmd, List<String> baseArgs) {
-			inactivityTimeout = Long.getLong(ENV_EXIF_TOOL_PROCESSCLEANUPDELAY,
-					DEFAULT_PROCESS_CLEANUP_DELAY);
-			startupArgs = new ArrayList<String>(baseArgs.size() + 5);
-			startupArgs.add(exifCmd);
-			startupArgs.addAll(Arrays.asList("-stay_open", "True"));
-			startupArgs.addAll(baseArgs);
-			startupArgs.addAll(Arrays.asList("-@", "-"));
-			// runs every minute to check if process has been inactive too long
-			cleanupTimer.schedule(new TimerTask() {
-				@Override
-				public void run() {
-					if (process != null && lastRunStart > 0
-							&& inactivityTimeout > 0) {
-						if ((System.currentTimeMillis() - lastRunStart) > inactivityTimeout) {
-							synchronized (this) {
-								if (process != null) {
-									process.close();
-								}
-							}
-						}
-					}
-				}
-			}, 60 * 1000); // every minute
-		}
-
-		public void setInactiveTimeout(long mills) {
-			this.inactivityTimeout = mills;
-		}
-
-		@Override
-		public void startup() {
-			shuttingDown.set(false);
-			if (process == null || process.isClosed()) {
-				synchronized (this) {
-					if (process == null || process.isClosed()) {
-						log.debug("Starting daemon ExifTool process and creating read/write streams (this only happens once)...");
-						process = new ExifProcess(true, startupArgs);
-					}
-				}
-			}
-		}
-
-		@Override
-		public Map<String, String> execute(final long runTimeoutMills,
-				List<String> args) throws IOException {
-			lastRunStart = System.currentTimeMillis();
-			int attempts = 0;
-			while (attempts < 3 && !shuttingDown.get()) {
-				attempts++;
-				if (process == null || process.isClosed()) {
-					synchronized (this) {
-						if (process == null || process.isClosed()) {
-							log.debug("Starting daemon ExifTool process and creating read/write streams (this only happens once)...");
-							process = new ExifProcess(true, startupArgs);
-						}
-					}
-				}
-				TimerTask attemptTimer = null;
-				try {
-					if (runTimeoutMills > 0) {
-						attemptTimer = new TimerTask() {
-							@Override
-							public void run() {
-								if (process != null && !process.isClosed()) {
-									log.warn("Process ran too long closing, max "
-											+ runTimeoutMills + " mills");
-									process.close();
-								}
-							}
-						};
-						cleanupTimer.schedule(attemptTimer, runTimeoutMills);
-					}
-					log.debug("Streaming arguments to ExifTool process...");
-					return process.sendToRunning(args);
-				} catch (IOException ex) {
-					if (STREAM_CLOSED_MESSAGE.equals(ex.getMessage())
-							&& !shuttingDown.get()) {
-						// only catch "Stream Closed" error (happens when
-						// process has died)
-						log.warn(String
-								.format("Caught IOException(\"%s\"), will restart daemon",
-										STREAM_CLOSED_MESSAGE));
-						process.close();
-					} else {
-						throw ex;
-					}
-				} finally {
-					if (attemptTimer != null)
-						attemptTimer.cancel();
-				}
-			}
-			if (shuttingDown.get()) {
-				throw new IOException("Shutting Down");
-			}
-			throw new IOException("Ran out of attempts");
-		}
-
-		@Override
-		public boolean isRunning() {
-			return process != null && !process.isClosed();
-		}
-
-		@Override
-		public void shutdown() {
-			shuttingDown.set(true);
-		}
-	}
-
-	public static class SingleUseExifProxy implements ExifProxy {
-		private final Timer cleanupTimer = new Timer(CLEANUP_THREAD_NAME, true);
-		private final List<String> baseArgs;
-
-		public SingleUseExifProxy(String exifCmd, List<String> defaultArgs) {
-			this.baseArgs = new ArrayList<String>(defaultArgs.size() + 1);
-			this.baseArgs.add(exifCmd);
-			this.baseArgs.addAll(defaultArgs);
-		}
-
-		@Override
-		public Map<String, String> execute(final long runTimeoutMills,
-				List<String> args) throws IOException {
-			List<String> newArgs = new ArrayList<String>(baseArgs.size()
-					+ args.size());
-			newArgs.addAll(baseArgs);
-			newArgs.addAll(args);
-			final ExifProcess process = new ExifProcess(false, newArgs);
-			TimerTask attemptTimer = null;
-			if (runTimeoutMills > 0) {
-				attemptTimer = new TimerTask() {
-					@Override
-					public void run() {
-						if (!process.isClosed()) {
-							log.warn("Process ran too long closing, max "
-									+ runTimeoutMills + " mills");
-							process.close();
-						}
-					}
-				};
-				cleanupTimer.schedule(attemptTimer, runTimeoutMills);
-			}
-			try {
-				return process.readResponse();
-			} finally {
-				process.close();
-				if (attemptTimer != null)
-					attemptTimer.cancel();
-			}
-		}
-
-		@Override
-		public void startup() {
-			;
-		}
-
-		@Override
-		public boolean isRunning() {
-			return false;
-		}
-
-		@Override
-		public void shutdown() {
-			;
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * All the read options, is immutable, copy on change, fluent style "with"
-	 * setters.
-	 */
-	public static class ReadOptions {
-		private final long runTimeoutMills;
-		private final boolean convertTypes;
-		private final boolean numericOutput;
-		private final boolean showDuplicates;
-		private final boolean showEmptyTags;
-
-		public ReadOptions() {
-			this(0, false, false, false, false);
-		}
-
-		private ReadOptions(long runTimeoutMills, boolean convertTypes,
-				boolean numericOutput, boolean showDuplicates,
-				boolean showEmptyTags) {
-			this.runTimeoutMills = runTimeoutMills;
-			this.convertTypes = convertTypes;
-			this.numericOutput = numericOutput;
-			this.showDuplicates = showDuplicates;
-			this.showEmptyTags = showEmptyTags;
-		}
-
-		@Override
-		public String toString() {
-			return String
-					.format("%s(runTimeout:%,d convertTypes:%s showDuplicates:%s showEmptyTags:%s)",
-							getClass().getSimpleName(), runTimeoutMills,
-							convertTypes, showDuplicates, showEmptyTags);
-		}
-
-		/**
-		 * Sets the maximum time a process can run
-		 */
-		public ReadOptions withRunTimeoutMills(long mills) {
-			return new ReadOptions(mills, convertTypes, numericOutput,
-					showDuplicates, showEmptyTags);
-		}
-
-		/**
-		 * By default all values will be returned as the strings printed by the
-		 * exiftool. If this is enabled then {@link MetadataTag#getType()} is
-		 * used to cast the string into a java type.
-		 */
-		public ReadOptions withConvertTypes(boolean enabled) {
-			return new ReadOptions(runTimeoutMills, enabled, numericOutput,
-					showDuplicates, showEmptyTags);
-		}
-
-		/**
-		 * Setting this to true will add the "-n" option causing the ExifTool to
-		 * output of some tags to change.
-		 * <p/>
-		 * ExifTool, via the <code>-n</code> command line arg, is capable of
-		 * returning most values in their raw numeric form (e.g.
-		 * Aperture="2.8010323841") as well as a more human-readable/friendly
-		 * format (e.g. Aperture="2.8").
-		 * <p/>
-		 * While the {@link Tag}s defined on this class do provide a hint at the
-		 * type of the result (see {@link MetadataTag#getType()}), that hint
-		 * only applies when the numeric value is returned.
-		 * <p/>
-		 * If the caller finds the human-readable format easier to process, Set
-		 * this to false, the default.
-		 * <p/>
-		 * In order to see the types of values that are returned when human
-		 * readable is used (default), you can check the comprehensive <a href=
-		 * "http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html">
-		 * ExifTool Tag Guide</a>.
-		 * <p/>
-		 * This makes sense with some values like Aperture that in numeric
-		 * format end up returning as 14-decimal-place, high precision values
-		 * that are near the intended value (e.g. "2.79999992203711" instead of
-		 * just returning "2.8"). On the other hand, other values (like
-		 * Orientation) are easier to parse when their numeric value (1-8) is
-		 * returned instead of a much longer friendly name (e.g.
-		 * "Mirror horizontal and rotate 270 CW").
-		 * 
-		 * Attempted from work done by
-		 * 
-		 * @author Riyad Kalla (software@thebuzzmedia.com)
-		 */
-		public ReadOptions withNumericOutput(boolean enabled) {
-			return new ReadOptions(runTimeoutMills, convertTypes, enabled,
-					showDuplicates, showEmptyTags);
-		}
-
-		/**
-		 * If enabled will show tags which are duplicated between different tag
-		 * regions, relates to the "-a" option in ExifTool.
-		 */
-		public ReadOptions withShowDuplicates(boolean enabled) {
-			return new ReadOptions(runTimeoutMills, convertTypes,
-					numericOutput, enabled, showEmptyTags);
-		}
-
-		public ReadOptions withShowEmptyTags(boolean enabled) {
-			return new ReadOptions(runTimeoutMills, convertTypes,
-					numericOutput, showDuplicates, enabled);
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * All the write options, is immutable, copy on change, fluent style "with"
-	 * setters.
-	 */
-	public static class WriteOptions {
-		private final long runTimeoutMills;
-		private final boolean deleteBackupFile;
-
-		public WriteOptions() {
-			this(0, false);
-		}
-
-		private WriteOptions(long runTimeoutMills, boolean deleteBackupFile) {
-			this.runTimeoutMills = runTimeoutMills;
-			this.deleteBackupFile = deleteBackupFile;
-		}
-
-		@Override
-		public String toString() {
-			return String.format("%s(runTimeOut:%,d deleteBackupFile:%s)",
-					getClass().getSimpleName(), runTimeoutMills,
-					deleteBackupFile);
-		}
-
-		public WriteOptions withRunTimeoutMills(long mills) {
-			return new WriteOptions(mills, deleteBackupFile);
-		}
-
-		/**
-		 * ExifTool automatically makes a backup copy a file before writing
-		 * metadata tags in the form "file.ext_original", by default this tool
-		 * will delete that original file after the writing is done.
-		 */
-		public WriteOptions withDeleteBackupFile(boolean enabled) {
-			return new WriteOptions(runTimeoutMills, enabled);
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Enum used to define the different kinds of features in the native
-	 * ExifTool executable that this class can help you take advantage of.
-	 * <p/>
-	 * These flags are different from {@link Tag}s in that a "feature" is
-	 * determined to be a special functionality of the underlying ExifTool
-	 * executable that requires a different code-path in this class to take
-	 * advantage of; for example, <code>-stay_open True</code> support.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Feature {
-		/**
-		 * Enum used to specify that you wish to launch the underlying ExifTool
-		 * process with <code>-stay_open True</code> support turned on that this
-		 * class can then take advantage of.
-		 * <p/>
-		 * Required ExifTool version is <code>8.36</code> or higher.
-		 */
-		STAY_OPEN(8, 36),
-		/**
-		 * Acitves the MWG modules. The Metadata Working Group (MWG) recommends
-		 * techniques to allow certain overlapping EXIF, IPTC and XMP tags to be
-		 * reconciled when reading, and synchronized when writing. The MWG
-		 * Composite tags below are designed to aid in the implementation of
-		 * these recommendations. Will add the args " -use MWG"
-		 * 
-		 * @see <a
-		 *      href="http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/MWG.html">ExifTool
-		 *      MWG Docs</a> !! Note these version numbers are not correct
-		 */
-		MWG_MODULE(8, 36), ;
-
-		private VersionNumber requireVersion;
-
-		private Feature(int... numbers) {
-			this.requireVersion = new VersionNumber(numbers);
-		}
-
-		/**
-		 * Used to get the version of ExifTool required by this feature in order
-		 * to work.
-		 * 
-		 * @return the version of ExifTool required by this feature in order to
-		 *         work.
-		 */
-		VersionNumber getVersion() {
-			return requireVersion;
-		}
-
-		boolean isSupported(VersionNumber exifVersionNumber) {
-			return requireVersion.isBeforeOrEqualTo(exifVersionNumber);
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Version Number used to determine if one version is after another.
-	 * 
-	 * @author Matt Gile, msgile
-	 */
-	static class VersionNumber {
-		private final int[] numbers;
-
-		public VersionNumber(String str) {
-			String[] versionParts = str.trim().split("\\.");
-			this.numbers = new int[versionParts.length];
-			for (int i = 0; i < versionParts.length; i++) {
-				numbers[i] = Integer.parseInt(versionParts[i]);
-			}
-		}
-
-		public VersionNumber(int... numbers) {
-			this.numbers = numbers;
-		}
-
-		public boolean isBeforeOrEqualTo(VersionNumber other) {
-			int max = Math.min(this.numbers.length, other.numbers.length);
-			for (int i = 0; i < max; i++) {
-				if (this.numbers[i] > other.numbers[i]) {
-					return false;
-				} else if (this.numbers[i] < other.numbers[i]) {
-					return true;
-				}
-			}
-			// assume missing number as zero, so if the current process number
-			// is more digits it is a higher version
-			return this.numbers.length <= other.numbers.length;
-		}
-
-		@Override
-		public String toString() {
-			StringBuilder builder = new StringBuilder();
-			for (int number : numbers) {
-				if (builder.length() > 0) {
-					builder.append(".");
-				}
-				builder.append(number);
-			}
-			return builder.toString();
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Base type for all "tag" passed to exiftool. The key is the value passed
-	 * to the exiftool like "-creator". The Types is used for automatic type
-	 * conversions.
-	 * 
-	 */
-	public interface MetadataTag {
-		/**
-		 * Returns the values passed to exiftool
-		 */
-		public String getKey();
-
-		/**
-		 * The types
-		 */
-		public Class getType();
-
-		public boolean isMapped();
-	}
-
-	// ================================================================================
-	/**
-	 * Enum used to pre-define a convenient list of tags that can be easily
-	 * extracted from images using this class with an external install of
-	 * ExifTool.
-	 * <p/>
-	 * Each tag defined also includes a type hint for the parsed value
-	 * associated with it when the default humanReadable is false.
-	 * <p/>
-	 * The types provided by each tag are merely a hint based on the <a
-	 * href="http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html"
-	 * >ExifTool Tag Guide</a> by Phil Harvey; the caller is free to parse or
-	 * process the returned {@link String} values any way they wish.
-	 * <h3>Tag Support</h3>
-	 * ExifTool is capable of parsing almost every tag known to man (1000+), but
-	 * this class makes an attempt at pre-defining a convenient list of the most
-	 * common tags for use.
-	 * <p/>
-	 * This list was determined by looking at the common metadata tag values
-	 * written to images by popular mobile devices (iPhone, Android) as well as
-	 * cameras like simple point and shoots as well as DSLRs. As an additional
-	 * source of input the list of supported/common EXIF formats that Flickr
-	 * supports was also reviewed to ensure the most common/useful tags were
-	 * being covered here.
-	 * <p/>
-	 * Please email me or <a
-	 * href="https://github.com/thebuzzmedia/imgscalr/issues">file an issue</a>
-	 * if you think this list is missing a commonly used tag that should be
-	 * added to it.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Tag implements MetadataTag {
-		// single entry tags
-		APERTURE("ApertureValue", Double.class),
-		AUTHOR("XPAuthor", String.class),
-		COLOR_SPACE("ColorSpace", Integer.class),
-		COMMENT("XPComment", String.class),
-		CONTRAST("Contrast", Integer.class),
-		CREATE_DATE("CreateDate", Date.class),
-		CREATION_DATE("CreationDate", Date.class),
-		DATE_CREATED("DateCreated", Date.class),
-		DATE_TIME_ORIGINAL("DateTimeOriginal", Date.class),
-		DIGITAL_ZOOM_RATIO("DigitalZoomRatio", Double.class),
-		EXIF_VERSION("ExifVersion", String.class),
-		EXPOSURE_COMPENSATION("ExposureCompensation", Double.class),
-		EXPOSURE_PROGRAM("ExposureProgram", Integer.class),
-		EXPOSURE_TIME("ExposureTime", Double.class),
-		FLASH("Flash", Integer.class),
-		FOCAL_LENGTH("FocalLength", Double.class),
-		FOCAL_LENGTH_35MM("FocalLengthIn35mmFormat", Integer.class),
-		FNUMBER("FNumber", String.class),
-		GPS_ALTITUDE("GPSAltitude", Double.class),
-		GPS_ALTITUDE_REF("GPSAltitudeRef", Integer.class),
-		GPS_BEARING("GPSDestBearing", Double.class),
-		GPS_BEARING_REF("GPSDestBearingRef", String.class),
-		GPS_DATESTAMP("GPSDateStamp", String.class),
-		GPS_LATITUDE("GPSLatitude", Double.class),
-		GPS_LATITUDE_REF("GPSLatitudeRef", String.class),
-		GPS_LONGITUDE("GPSLongitude", Double.class),
-		GPS_LONGITUDE_REF("GPSLongitudeRef", String.class),
-		GPS_PROCESS_METHOD("GPSProcessingMethod", String.class),
-		GPS_SPEED("GPSSpeed", Double.class),
-		GPS_SPEED_REF("GPSSpeedRef", String.class),
-		GPS_TIMESTAMP("GPSTimeStamp", String.class),
-		IMAGE_HEIGHT("ImageHeight", Integer.class),
-		IMAGE_WIDTH("ImageWidth", Integer.class),
-		ISO("ISO", Integer.class),
-		KEYWORDS("XPKeywords", String.class),
-		LENS_MAKE("LensMake", String.class),
-		LENS_MODEL("LensModel", String.class),
-		MAKE("Make", String.class),
-		METERING_MODE("MeteringMode", Integer.class),
-		MODEL("Model", String.class),
-		ORIENTATION("Orientation", Integer.class),
-		OWNER_NAME("OwnerName", String.class),
-		RATING("Rating", Integer.class),
-		RATING_PERCENT("RatingPercent", Integer.class),
-		ROTATION("Rotation", Integer.class),
-		SATURATION("Saturation", Integer.class),
-		SENSING_METHOD("SensingMethod", Integer.class),
-		SHARPNESS("Sharpness", Integer.class),
-		SHUTTER_SPEED("ShutterSpeedValue", Double.class),
-		SOFTWARE("Software", String.class),
-		SUBJECT("XPSubject", String.class),
-		TITLE("XPTitle", String.class),
-		WHITE_BALANCE("WhiteBalance", Integer.class),
-		X_RESOLUTION("XResolution", Double.class),
-		Y_RESOLUTION("YResolution", Double.class), ;
-
-		/**
-		 * Used to get the {@link Tag} identified by the given, case-sensitive,
-		 * tag name.
-		 * 
-		 * @param name
-		 *            The case-sensitive name of the tag that will be searched
-		 *            for.
-		 * 
-		 * @return the {@link Tag} identified by the given, case-sensitive, tag
-		 *         name or <code>null</code> if one couldn't be found.
-		 */
-		public static Tag forName(String name) {
-			for (Tag tag : Tag.values()) {
-				if (tag.getKey().equals(name)) {
-					return tag;
-				}
-			}
-			return null;
-		}
-
-		public static Map<Tag, String> toTagMap(Map<String, String> values) {
-			Map<Tag, String> tagMap = new EnumMap<Tag, String>(Tag.class);
-			for (Tag tag : Tag.values()) {
-				if (values.containsKey(tag.getKey())) {
-					tagMap.put(tag, values.get(tag.getKey()));
-				}
-			}
-			return tagMap;
-
-		}
-
-		/**
-		 * Convenience method used to convert the given string Tag value
-		 * (returned from the external ExifTool process) into the type described
-		 * by the associated {@link Tag}.
-		 * 
-		 * @param <T>
-		 *            The type of the returned value.
-		 * @param value
-		 *            The {@link String} representation of the tag's value as
-		 *            parsed from the image.
-		 * 
-		 * @return the given string value converted to a native Java type (e.g.
-		 *         Integer, Double, etc.).
-		 * 
-		 * @throws IllegalArgumentException
-		 *             if <code>tag</code> is <code>null</code>.
-		 * @throws NumberFormatException
-		 *             if any exception occurs while trying to parse the given
-		 *             <code>value</code> to any of the supported numeric types
-		 *             in Java via calls to the respective <code>parseXXX</code>
-		 *             methods defined on all the numeric wrapper classes (e.g.
-		 *             {@link Integer#parseInt(String)} ,
-		 *             {@link Double#parseDouble(String)} and so on).
-		 * @throws ClassCastException
-		 *             if the type defined by <code>T</code> is incompatible
-		 *             with the type defined by {@link Tag#getType()} returned
-		 *             by the <code>tag</code> argument passed in. This class
-		 *             performs an implicit/unchecked cast to the type
-		 *             <code>T</code> before returning the parsed result of the
-		 *             type indicated by {@link Tag#getType()}. If the types do
-		 *             not match, a <code>ClassCastException</code> will be
-		 *             generated by the VM.
-		 */
-		@SuppressWarnings("unchecked")
-		public <T> T parseValue(String value) throws IllegalArgumentException {
-			return (T) deserialize(getKey(), value, getType());
-		}
-
-		/**
-		 * Used to get the name of the tag (e.g. "Orientation", "ISO", etc.).
-		 * 
-		 * @return the name of the tag (e.g. "Orientation", "ISO", etc.).
-		 */
-		public String getName() {
-			return name;
-		}
-
-		/**
-		 * Used to get a hint for the native type of this tag's value as
-		 * specified by Phil Harvey's <a href=
-		 * "http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html"
-		 * >ExifTool Tag Guide</a>.
-		 * 
-		 * @return a hint for the native type of this tag's value.
-		 */
-		@Override
-		public Class<?> getType() {
-			return type;
-		}
-
-		@Override
-		public String getKey() {
-			return name;
-		}
-
-		@Override
-		public boolean isMapped() {
-			return true;
-		}
-
-		private String name;
-		private Class<?> type;
-
-		private Tag(String name, Class<?> type) {
-			this.name = name;
-			this.type = type;
-		}
-	}
-
-	// ================================================================================
-	public enum MwgTag implements MetadataTag {
-		LOCATION("Location", String.class),
-		CITY("City", String.class),
-		STATE("State", String.class),
-		COUNTRY("Country", String.class),
-
-		COPYRIGHT("Copyright", String.class),
-
-		DATE_TIME_ORIGINAL("DateTimeOriginal", Date.class),
-		CREATE_DATE("CreateDate", Date.class),
-		MODIFY_DATE("ModifyDate", Date.class),
-
-		CREATOR("Creator", String.class),
-		DESCRIPTION("Description", String.class),
-		KEYWORDS("Keywords", String[].class),
-
-		ORIENTATION("Orientation", Integer.class),
-		RATING("Rating", Integer.class),
-
-		;
-
-		private String name;
-		private Class<?> type;
-
-		private MwgTag(String name, Class type) {
-			this.name = name;
-			this.type = type;
-		}
-
-		@Override
-		public String getKey() {
-			return name;
-		}
-
-		@Override
-		public Class<?> getType() {
-			return type;
-		}
-
-		@Override
-		public boolean isMapped() {
-			return true;
-		}
-
-		@Override
-		public String toString() {
-			return name;
-		}
-	}
-
-	// ================================================================================
-
-	/**
-	 * A Custom Tag that the user defines. Used to cover tags not in the enum.
-	 */
-	public static class CustomTag implements MetadataTag {
-		private final String name;
-		private final Class type;
-		private final boolean mapped;
-
-		public CustomTag(String name, Class type) {
-			this(name, type, !name.trim().endsWith(":all"));
-		}
-
-		public CustomTag(String name, Class type, boolean mapped) {
-			this.name = name.trim();
-			this.type = type;
-			this.mapped = mapped;
-		}
-
-		@Override
-		public String getKey() {
-			return name;
-		}
-
-		@Override
-		public Class getType() {
-			return type;
-		}
-
-		@Override
-		public boolean isMapped() {
-			return mapped;
-		}
-
-		@Override
-		public String toString() {
-			return getKey();
-		}
-	}
-
-	// ================================================================================
-	public enum TagGroup implements MetadataTag {
-		EXIF("EXIF", "exif:all"),
-		IPTC("IPTC", "iptc:all"),
-		XMP("XMP", "xmp:all"),
-		ALL("ALL", "all");
-
-		private final String name;
-		private final String key;
-
-		private TagGroup(String name, String key) {
-			this.name = name;
-			this.key = key;
-		}
-
-		public String getName() {
-			return name;
-		}
-
-		@Override
-		public String getKey() {
-			return key;
-		}
-
-		@Override
-		public Class getType() {
-			return Void.class;
-		}
-
-		@Override
-		public boolean isMapped() {
-			return false;
-		}
-
-	}
-
-	// ================================================================================
-	/**
-	 * Class used to define an exception that occurs when the caller attempts to
-	 * use a {@link Feature} that the underlying native ExifTool install does
-	 * not support (i.e. the version isn't new enough).
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public class UnsupportedFeatureException extends RuntimeException {
-		private static final long serialVersionUID = -1332725983656030770L;
-
-		public UnsupportedFeatureException(Feature feature) {
-			super(
-					"Use of feature ["
-							+ feature.toString()
-							+ "] requires version "
-							+ feature.getVersion()
-							+ " or higher of the native ExifTool program. The version of ExifTool referenced by the system property 'exiftool.path' is not high enough. You can either upgrade the install of ExifTool or avoid using this feature to workaround this exception.");
-		}
-	}
 }
