+++ src/main/java/com/thebuzzmedia/exiftool/Format.java
@@ -1,0 +1,8 @@
+package com.thebuzzmedia.exiftool;
+
+// ================================================================================
+// OLD API
+public enum Format {
+	NUMERIC,
+	HUMAN_READABLE
+}
