+++ src/main/java/com/thebuzzmedia/exiftool/TagGroup.java
@@ -1,0 +1,37 @@
+package com.thebuzzmedia.exiftool;
+
+// ================================================================================
+public enum TagGroup implements MetadataTag {
+	EXIF("EXIF", "exif:all"),
+	IPTC("IPTC", "iptc:all"),
+	XMP("XMP", "xmp:all"),
+	ALL("ALL", "all");
+
+	private final String name;
+	private final String key;
+
+	private TagGroup(String name, String key) {
+		this.name = name;
+		this.key = key;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	@Override
+	public String getKey() {
+		return key;
+	}
+
+	@Override
+	public Class getType() {
+		return Void.class;
+	}
+
+	@Override
+	public boolean isMapped() {
+		return false;
+	}
+
+}
